added fragment homework

+++ 04-retrofit/app/src/main/java/com/example/a04_retrofit/models/Coord.kt
@@ -1,6 +1,6 @@
 package com.example.a04_retrofit.models
 
 data class Coord(
-    val lat: Int,
-    val lon: Int
+    val lat: Float,
+    val lon: Float
 )
